--- constants/platformVersion.ts
@@ -1,6 +1,0 @@
-export const PLATFORM_VERSIONS = {
-  v2023_2: '2023.2',
-  v2025_1: '2025.1',
-  v2025_2: '2025.2',
-  unstable: 'unstable',
-};
